Limiting mouse movement, crosshair addition
I added a script where it would restrict the movement of the player, since there was an error or I missed a part of the code. But the summary is that the script stops the player from infinitely turning around down the y axis which if not fixed would've caused a weird perspective for the user.

I have also added a crosshair for the user, being a small dot, maybe I will change it so that there are more options to choose from but I am not sure yet.